==> analysis/2_analysis.qmd <==
---
title: "FaceNewsIntervention - Data Analysis"
editor: source
editor_options: 
  chunk_output_type: console
format:
  html:
    code-fold: true
    self-contained: false
    toc: true
execute:
  cache: true
---


## Data Preparation

```{r}
#| message: false
#| warning: false

library(tidyverse)
library(easystats)
library(patchwork)
library(ggside)
library(glmmTMB)
```


```{r}
#| code-fold: false

df <- read.csv("../data/data_participants.csv") |> 
  mutate(
    # Political_Affiliation = fct_relevel(Political_Affiliation, "Green", "Labour", "Democrat", "None", "None or Independent", "Liberal-Democrats", "Conservative", "Republican"),
    Political_Ideology = fct_relevel(Political_Ideology, "Extremely Liberal", "Liberal", "Slightly Liberal", "Moderate", "Don't Know", "Slightly Conservative", "Conservative", "Extremely Conservative"))

dfmist <- read.csv("../data/data_mist.csv") |> 
  mutate(
         Condition = ifelse(Condition == "Pretest", "Pre-test", "Post-test"),
         Condition = fct_relevel(Condition, "Pre-test", "Post-test"),
         Intervention = ifelse(Intervention == "Tetris", "Tetris", "Bad News"),
         Intervention = fct_relevel(Intervention, "Bad News", "Tetris"),
         Topic = ifelse(Topic == "Covid", "Covid items", "General items"),
         Topic = fct_relevel(Topic, "General items", "Covid items"),
         Belief = ifelse(Realness > 0.5, 0, 1),
         Fakeness = 1 - Realness,
         Confidence = abs(datawizard::rescale(Fakeness, to = c(-1, 1), range = c(0, 1)))) |> 
  full_join(select(df, Participant, Sample, Intervention_Valid, Political_IdeologyNumeric, starts_with("BFI"), starts_with("GCBS"), starts_with("VSA"), COVID_Vaccination, Initial_SubjectiveDiscrimination, Intervention_SubjectiveEffect), by = "Participant") |> 
  filter(Sample == "USA")

dfmocri <- read.csv("../data/data_mocri.csv") |> 
  full_join(select(df, Participant, Sample, Intervention, Intervention_Valid, Political_IdeologyNumeric, starts_with("BFI"), starts_with("GCBS"), starts_with("VSA"), COVID_Vaccination, Initial_SubjectiveDiscrimination, Intervention_SubjectiveEffect), by = "Participant") |> 
  mutate(Condition = fct_relevel(Condition, "Pre-test", "Post-test"),
         Intervention = ifelse(Intervention == "Tetris", "Tetris", "Bad News"),
         Intervention = fct_relevel(Intervention, "Bad News", "Tetris")) |> 
  filter(Sample == "USA")
```

## Intervention 

### MIST - Beliefs

```{r}
#| code-fold: false

dfmist[dfmist$Intervention_Valid == TRUE, ] |> 
  ggplot(aes(x=Realness)) +
  geom_histogram(aes(fill=Nature), alpha=0.6, position="identity", bins=30) +
  theme_minimal()  +
  scale_fill_manual(values=c("Fake"="#F44336", "Real"="#4CAF50"))

m1 <- glmmTMB(
  Belief ~ Intervention * Condition * Nature * Topic + (Condition * Nature|Participant) + (1|Item ),
  data = dfmist[dfmist$Intervention_Valid == TRUE, ],
  family = "binomial"
)

means1 <- estimate_means(m1, by=c("Nature", "Condition", "Intervention", "Topic"))
means1$Topic <- paste0("MIST (", means1$Topic, ")")



display_contrasts <- function(m) {
  
  if("Topic" %in% insight::find_predictors(m)$conditional) {
    by <- c("Topic", "Nature")
  } else {
    by <- c("Nature")
  }
  
  c1 <- estimate_contrasts(m, contrast="Condition", by=c(by, "Intervention"), p_adjust="holm") |> 
    select(-SE, -z) |> 
    insight::format_table() |> 
    gt::gt() |> 
    gt::data_color(columns = "Nature", palette = c("#F44336", "#4CAF50")) |> 
    gt::data_color(columns = "Intervention", palette = c("purple", "yellow"))
  
  c2 <- estimate_contrasts(m, contrast="Intervention", by=c(by, "Condition='Pre-test'"), p_adjust="holm") |> 
    select(-SE, -z) |>  
    insight::format_table() |> 
    gt::gt() |> 
    gt::data_color(columns = "Nature", palette = c("#F44336", "#4CAF50"))
  
  list(c1=c1, c2=c2)
}

c <- display_contrasts(m1)
c$c1
c$c2
```

### New Model for intervention effect by political ideology

```{r}
# m9 <- glmmTMB(
#   Belief ~ Intervention * Condition * Nature * Topic * Political_IdeologyNumeric  + (Condition * Nature|Participant) + (1|Item ),
#   data = dfmist[dfmist$Intervention_Valid == TRUE, ],
#   family = "binomial"
# )
# 
# saveRDS(m9, "m9_model.rds")
# 
# m9 <- readRDS("m9_model.rds")
# 
# 
# means9 <- estimate_means(
#   m9,
#   at = list(Political_IdeologyNumeric = seq(1, 7, by = 0.5)),
#   by = c("Political_IdeologyNumeric", "Condition", "Intervention")
# )
# 
# saveRDS(means9, file = "means9.rds")
# means9 <- readRDS("means9.rds")
# 
# 
# p9 <- ggplot(means9, aes(x = Political_IdeologyNumeric, y = Probability, 
#                          color = Nature, linetype = Intervention)) +
#   geom_hline(yintercept = 0.5, color = "lightgrey") +
#   geom_line(linewidth = 1) +
#   geom_ribbon(aes(ymin = CI_low, ymax = CI_high, fill = Nature), alpha = 0.2, color = NA) +
#   scale_color_manual(values = c("Fake" = "#F44336", "Real" = "#4CAF50")) +
#   scale_fill_manual(values = c("Fake" = "#F44336", "Real" = "#4CAF50")) +
#   scale_y_continuous(labels = scales::percent) +
#   theme_minimal() +
#   labs(x = "Political Ideology (1 = Very Liberal, 7 = Very Conservative)",
#        y = "Probability of Believing an Item Fake/Manipulative",
#        color = "Nature", linetype = "Intervention") +
#   facet_wrap(~Topic) +
#   theme(strip.background = element_rect(fill = "lightgrey", color = "white"))
# 
# p9

```

### New linear model for intervention effect by political ideology

```{r}
df$MIST_correct_Diff <- df$MIST_Correct_Posttest - df$MIST_Correct_Pretest

df_clean <- df %>%
  filter(!is.na(Political_IdeologyNumeric), Intervention == "BadNewsGame") 


model <- lm(MIST_correct_Diff ~ Political_IdeologyNumeric, data = df_clean)
summary(model)


ggplot(df_clean, aes(x = Political_IdeologyNumeric, y = MIST_correct_Diff)) +
  geom_jitter(width = 0.2, alpha = 0.6) +        # spread points horizontally
  stat_summary(fun = mean, geom = "point", color = "red", size = 3) +  # mean points
  stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red", width = 0.2) + # CIs
  geom_smooth(method = "lm", se = TRUE, color = "blue") +
  labs(
    x = "Political Ideology (Numeric)",
    y = "MIST Correct Post-Pre Difference",
    title = "Relationship between Political Ideology and MIST Difference"
  ) +
  theme_minimal()

model_quad <- lm(MIST_correct_Diff ~ Political_IdeologyNumeric + I(Political_IdeologyNumeric^2), data = df_clean)


ggplot(df_clean, aes(x = Political_IdeologyNumeric, y = MIST_correct_Diff)) +
  geom_jitter(width = 0.2, alpha = 0.6) +
  stat_summary(fun = mean, geom = "point", color = "red", size = 3) +
  stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red", width = 0.2) +
  geom_smooth(method = "lm", formula = y ~ x + I(x^2), se = TRUE, color = "blue") +
  labs(
    x = "Political Ideology (Numeric)",
    y = "MIST Correct Post-Pre Difference",
    title = "Quadratic Relationship between Political Ideology and MIST Difference"
  ) +
  theme_minimal()


```

### New linear model for intervention effect (covid fake only) by political ideology

```{r}
df_filtered <- dfmist%>%
  filter(Nature == "Fake",
         Topic == "Covid items",
         Intervention == "Bad News")

df_summary <- df_filtered %>%
  group_by(Participant, Condition) %>%
  summarise(
    correct_sum = sum(Correct),
    total_items = n(),
    prop_correct = correct_sum / total_items
  ) %>%
  ungroup()


df_summary_wide <- df_summary %>%
  select(Participant, Condition, prop_correct) %>%
  pivot_wider(names_from = Condition, values_from = prop_correct) %>%
  left_join(df %>% select(Participant, Political_IdeologyNumeric), by = "Participant") %>%
  mutate(prop_diff = `Post-test` - `Pre-test`)

model2 <- lm(prop_diff ~ Political_IdeologyNumeric, data = df_summary_wide)
summary(model2)

ggplot(df_summary_wide, aes(x = Political_IdeologyNumeric, y = prop_diff)) +
  geom_point(alpha = 0.6) +
  geom_smooth(method = "lm", se = TRUE, color = "blue") +
  labs(
    x = "Political Ideology (Numeric)",
    y = "Change in Proportion Correct (Post - Pre)",
    title = "Relationship between Political Ideology and Improvement"
  ) +
  theme_minimal()

```


```{r}
dfsub <- estimate_grouplevel(m1) 

left_join(dfsub, select(df, Level=Participant, Intervention), by = "Level") |>
  filter(Intervention == "BadNewsGame") |>
  performance::performance_dvour()

# dfsub |> 
#   filter(Group != "Item", Parameter %in% c("(Intercept)", "ConditionPre-test:NatureReal")) 
```


```{r}
# m <- glmmTMB(
#   Fakeness ~ Intervention / Condition / Nature * Topic + (Nature / Condition|Participant) + (1|Item ),
#   data = dfmist,
#   family = ordbeta()
# )
# 
# means <- estimate_means(m, by=c("Nature", "Condition", "Intervention", "Topic"))
# p_mist2 <- means |> 
#   ggplot(aes(x=Condition, y=Proportion, color=Nature, group=interaction(Nature, Intervention))) +
#   geom_line(aes(linetype=Intervention), position=position_dodge2(width=0.2)) +
#   geom_pointrange(aes(ymin=CI_low, ymax=CI_high), position=position_dodge2(width=0.2)) +
#   facet_wrap(~Topic) +
#   scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50")) +
#   scale_y_continuous(labels=scales::percent) + 
#   theme_minimal() + 
#   labs(y="Belief\nReal vs. Fake", title = "MIST") +
#   theme(axis.title.x=element_blank(),
#         strip.background = element_rect(fill="lightgrey", color="white"))
# p_mist2
# 
# # p_mist1 / p_mist2
# 
# estimate_contrasts(m, contrast="Condition", by=c("Topic", "Nature", "Intervention"), p_adjust="holm")
# estimate_contrasts(m, contrast="Intervention", by=c("Topic", "Nature", "Condition='Pre-test'"), p_adjust="holm")
```


### MIST - Confidence

```{r}
#| code-fold: false

dfmist[dfmist$Intervention_Valid == TRUE, ] |> 
  ggplot(aes(x=Confidence)) +
  geom_histogram(aes(fill=Nature), alpha=0.6, position="identity", bins=30) +
  theme_minimal()  +
  scale_fill_manual(values=c("Fake"="#F44336", "Real"="#4CAF50"))


m2 <- glmmTMB(
  Confidence ~ Intervention * Condition * Nature * Topic + (Nature * Condition|Participant) + (1|Item ),
  data = filter(dfmist, Correct == 1, Intervention_Valid == TRUE),
  family = ordbeta()
)

means2 <- estimate_means(m2, by=c("Nature", "Condition", "Intervention", "Topic"))
means2$Topic <- paste0("MIST (", means2$Topic, ")")

p_mist2 <- means2 |>
  ggplot(aes(x=Condition, y=Proportion, color=Nature, group=interaction(Nature, Intervention))) +
  geom_line(aes(linetype=Intervention), position=position_dodge2(width=0.2)) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high), position=position_dodge2(width=0.2)) +
  facet_wrap(~Topic) +
  scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50")) +
  scale_y_continuous(labels=scales::percent) +
  theme_minimal() +
  labs(y="Correct Responses - Confidence", color="Items") +
  theme(axis.title.x=element_blank(),
        strip.background = element_rect(fill="lightgrey", color="white"))
p_mist2
# 
# # p_mist1 / p_mist2

c <- display_contrasts(m2)
c$c1
c$c2
```


### MOCRI

```{r}
#| code-fold: false

m3 <- glmmTMB(
  Manipulative ~ Intervention * Nature * Condition + (Nature | Participant) + (1|Item ),
  data = dfmocri[dfmocri$Intervention_Valid == TRUE, ],
  family = "binomial"
)

means3 <- estimate_means(m3, by=c("Nature", "Condition", "Intervention"))
means3$Topic <- "MOCRI"

c <- display_contrasts(m3)
c$c1
c$c2
```

```{r}
p1 <- rbind(means1, means3) |> 
  ggplot(aes(x=Condition, y=Probability, color=Nature, group=interaction(Nature, Intervention))) +
  geom_hline(yintercept=0.5, color="lightgrey") +
  geom_line(aes(linetype=Intervention), position=position_dodge2(width=0.2), linewidth = 1) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high), position=position_dodge2(width=0.2), size = 1) +
  scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
  scale_y_continuous(labels=scales::percent) + 
  theme_minimal() + 
  labs(y="Probability of Believing an Item Fake/Manipulative", color="Items") +
  facet_wrap(~Topic) +
  theme(strip.background = element_rect(fill="lightgrey", color="white"),
        axis.title.x=element_blank()) +
  geom_text(data = rbind(
    data.frame(x = 1.48, y = 0.42, Nature="Fake", Intervention="Bad News", Topic = "MIST (Covid items)", label = c("*")),
    data.frame(x = 1.48, y = 0.7, Nature="Fake", Intervention="Bad News", Topic = "MIST (General items)", label = c("**")),
    data.frame(x = 0.98, y = 0.69, Nature="Fake", Intervention="Bad News", Topic = "MIST (General items)", label = c("**")),
    data.frame(x = 0.98, y = 0.75, Nature="Fake", Intervention="Bad News", Topic = "MOCRI", label = c(""))
  ),
    aes(x = x, y = y, label = label)
  )
p1
```


### Subjective Effect

```{r}
#| code-fold: false

#MIST
m1 <- glmmTMB(
  Belief ~ Intervention_SubjectiveEffect * Intervention * Condition * Nature * Topic + (Condition * Nature|Participant) + (1|Item ),
  data = dfmist[dfmist$Intervention_Valid == TRUE, ],
  family = "binomial"
)

c <- estimate_contrasts(m1, contrast="Condition", by=c("Intervention_SubjectiveEffect", "Topic", "Nature", "Intervention"), p_adjust="holm")
c


#MIST conf
m2 <- glmmTMB(
  Confidence ~ Intervention_SubjectiveEffect * Intervention * Condition * Nature * Topic + (Nature * Condition|Participant) + (1|Item ),
  data = filter(dfmist, Correct == 1, Intervention_Valid == TRUE),
  family = ordbeta()
)

means2 <- estimate_means(m2, by=c("Intervention_SubjectiveEffect", "Nature", "Condition", "Intervention", "Topic"))
means2$Topic <- paste0("MIST (", means2$Topic, ")")

p_mist2 <- means2 |>
  ggplot(aes(x=Condition, y=Proportion, color=Nature, group=interaction(Nature, Intervention))) +
  geom_line(aes(linetype=Intervention), position=position_dodge2(width=0.2)) +
  geom_pointrange(aes(ymin=CI_low, ymax=CI_high), position=position_dodge2(width=0.2)) +
  facet_wrap(~Topic) +
  scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50")) +
  scale_y_continuous(labels=scales::percent) +
  theme_minimal() +
  labs(y="Correct Responses - Confidence", color="Items") +
  theme(axis.title.x=element_blank(),
        strip.background = element_rect(fill="lightgrey", color="white"))
p_mist2
# 
# # p_mist1 / p_mist2

c <- display_contrasts(m2)
c$c1
c$c2


#MOCRI
m3 <- glmmTMB(
  Manipulative ~ Intervention_SubjectiveEffect * Intervention * Nature * Condition + (Nature | Participant) + (1|Item ),
  data = dfmocri[dfmocri$Intervention_Valid == TRUE, ],
  family = "binomial"
)

means3 <- estimate_means(m3, by=c("Intervention_SubjectiveEffect", "Nature", "Condition", "Intervention"))
means3$Topic <- "MOCRI"

c <- display_contrasts(m3)
c$c1
c$c2

# TODO: add conf and mocri models from above and add subjectiveeffect to the formula.
```

## Determinants

::: panel-tabset

### Political Ideology

```{r}
make_covariatemodel <- function(covariate="Political_IdeologyNumeric") {
  # MIST ---------------------------------------------------------------------
  f <- paste0("Belief ~ Nature / poly(", 
              covariate, 
              ", 2) * Topic + (Nature|Participant) + (1|Item)")
  
  m <- glmmTMB(
    as.formula(f),
    data = dfmist[!is.na(dfmist[[covariate]]) & dfmist$Condition == "Pre-test",],
    family = "binomial"
  )
  
  # m <- glmmTMB(
  #   Fakeness ~ Nature / poly(Political_IdeologyNumeric, 2) * Topic + (Nature|Participant) + (1|Item ),
  #   data = filter(dfmist, Condition == "Pre-test", !is.na(Political_IdeologyNumeric)),
  #   family = ordbeta()
  # )
  
  # parameters::parameters(m)

  
  pred <- estimate_relation(m, by=c("Nature", covariate, "Topic"), length = 20) 
  pred$x <- pred[[covariate]]
  pred$Topic <- paste0("MIST (", pred$Topic, ")")
  
  # Benchmark
  trends1 <- estimate_slopes(m, trend=covariate, by=c(covariate, "Nature", "Topic"), length=5, backend = "emmeans")
  
  # Confidence ----------------------------------------------------------------
  f <- paste0("Confidence ~ Nature / poly(", 
              covariate, 
              ", 2) * Topic + (Nature|Participant) + (1|Item)")
  
  m <- glmmTMB(
    as.formula(f),
    data = dfmist[!is.na(dfmist[[covariate]]) & dfmist$Condition == "Pre-test" & dfmist$Correct == 1,],
    family = ordbeta()
  )
  
  pred2 <- estimate_relation(m, by=c("Nature", covariate, "Topic"), length = 20) 
  pred2$x <- pred2[[covariate]]
  pred2$Topic <- paste0("MIST (", pred2$Topic, ")")
  
  p_conf <- pred2 |> 
    ggplot(aes(x=x, y=Predicted)) +
    geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Nature), alpha=0.3) +
    geom_line(aes(color=Nature)) +
    scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
    scale_fill_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
    scale_y_continuous(labels=scales::percent) + 
    theme_minimal() +
    facet_wrap(~Topic) +
    labs(x = covariate, y="Correct Responses - Confidence") +
    theme(strip.background = element_rect(fill="lightgrey", color="white"),
          axis.title.x = element_text(face="bold"))
  
  trends_conf <- estimate_slopes(m, trend=covariate, by=c(covariate, "Nature", "Topic"), length=5, backend="emmeans") |> 
    select(-SE, -z, -df_error) |> 
    insight::format_table() |> 
    gt::gt() |> 
    gt::data_color(columns = "Nature", palette = c("#F44336", "#4CAF50")) |> 
    gt::data_color(columns = covariate, palette = c("lightblue", "blue")) |> 
    gt::data_color(columns = "p", fn = function(x) ifelse(x < .01, "orange", ifelse(x < .05, "gold", "white")))
  
  # MOCRI --------------------------------------------------------------------
  f <- paste0("Manipulative ~ Nature / poly(", 
              covariate, 
              ", 2) + (Nature|Participant) + (1|Item)")
  
  m <- glmmTMB(
    as.formula(f),
    data = dfmocri[!is.na(dfmocri[[covariate]]) & dfmocri$Condition == "Pre-test",], 
    family = "binomial"
  )
  
  pred3 <- estimate_relation(m, by=c("Nature", covariate), length = 20) 
  pred3$x <- pred3[[covariate]]
  pred3$Topic <- "MOCRI"
  
  p <- rbind(pred, pred3) |> 
    ggplot(aes(x=x, y=Predicted)) +
    geom_ribbon(aes(ymin=CI_low, ymax=CI_high, fill=Nature), alpha=0.3) +
    geom_line(aes(color=Nature)) +
    scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
    scale_fill_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
    scale_y_continuous(labels=scales::percent) + 
    theme_minimal() +
    facet_wrap(~Topic) +
    labs(x = covariate, y = "Probability of Believing an Item Fake/Manipulative") +
    theme(strip.background = element_rect(fill="lightgrey", color="white"),
          axis.title.x = element_text(face="bold"))
  
  trends2 <- estimate_slopes(m, trend=covariate, by=c(covariate, "Nature"), length=5, backend="emmeans") 
  trends2$Topic <- "MOCRI"
  
  trends <- rbind(trends1, trends2) |> 
    select(-SE, -z, -df_error) |> 
    insight::format_table() |> 
    gt::gt() |> 
    gt::data_color(columns = "Nature", palette = c("#F44336", "#4CAF50")) |> 
    gt::data_color(columns = covariate, palette = c("lightblue", "blue")) |> 
    gt::data_color(columns = "p", fn = function(x) ifelse(x < .01, "orange", ifelse(x < .05, "gold", "white")))
  
  list(p=p, trends=trends, p_conf = p_conf, trends_conf=trends_conf)
}

rez_ideology <- make_covariatemodel(covariate="Political_IdeologyNumeric")
rez_ideology$p
rez_ideology$trends

rez_ideology$p_conf
rez_ideology$trends_conf
```

### Authoritarianism

```{r}
rez_vsa <- make_covariatemodel(covariate="VSA_General")
rez_vsa$p
rez_vsa$trends

rez_vsa$p_conf
rez_vsa$trends_conf
```

### Personality

```{r}
rez_neuro <- make_covariatemodel(covariate="BFI_Neuroticism")
rez_neuro$p
rez_neuro$trends

rez_neuro$p_conf
rez_neuro$trends_conf
```

```{r}
rez_consc <- make_covariatemodel(covariate="BFI_Conscientiousness")
rez_consc$p
rez_consc$trends

rez_consc$p_conf
rez_consc$trends_conf
```

```{r}
rez_agre <- make_covariatemodel(covariate="BFI_Agreeableness")
rez_agre$p
rez_agre$trends

rez_agre$p_conf
rez_agre$trends_conf
```

```{r}
rez_open <- make_covariatemodel(covariate="BFI_Openness")
rez_open$p
rez_open$trends 

rez_open$p_conf
rez_open$trends_conf
```

```{r}
rez_extr <- make_covariatemodel(covariate="BFI_Extraversion")
rez_extr$p
rez_extr$trends

rez_extr$p_conf
rez_extr$trends_conf 
```

### Conspiracy

```{r}
rez_gcbsg <- make_covariatemodel(covariate="GCBS_General")
rez_gcbsg$p
rez_gcbsg$trends |>
  filter(p < .05)

rez_gcbsg$p_conf
rez_gcbsg$trends_conf |>
  filter(p < .05)
```

```{r}
rez_gcbs2 <- make_covariatemodel(covariate="GCBS_GovernmentMalfeasance")
rez_gcbs2$p
rez_gcbs2$trends |>
  filter(p < .05)

rez_gcbs2$p_conf
rez_gcbs2$trends_conf |>
  filter(p < .05)
```


### Initial Fakenews Subjective Discrimination 

```{r}
# MIST 
m <- glmmTMB(
  Belief ~ Nature / Initial_SubjectiveDiscrimination * Topic + (Nature|Participant) + (1|Item),
  data = dfmist[!is.na(dfmist$COVID_Vaccination) & dfmist$Condition == "Pre-test",],
  family = "binomial"
)

contrasts_mist <- estimate_contrasts(m, contrast = "Initial_SubjectiveDiscrimination", by = c("Nature", "Topic"))

pred_mist <- estimate_means(m, by=c("Nature", "Initial_SubjectiveDiscrimination", "Topic")) 
pred_mist$Topic <- paste0("MIST (", pred_mist$Topic, ")")
  


m <- glmmTMB(
  Confidence ~ Nature / Initial_SubjectiveDiscrimination * Topic + (Nature|Participant) + (1|Item),
  data = dfmist[!is.na(dfmist$Initial_SubjectiveDiscrimination) & dfmist$Condition == "Pre-test" & dfmist$Correct == 1,],
  family = ordbeta()
)

contrasts_conf <- estimate_contrasts(m, contrast = "Initial_SubjectiveDiscrimination", by = c("Nature", "Topic"))
contrasts_conf$Topic <- paste0("Confidence: ", contrasts_conf$Topic, "")

pred_conf <- estimate_means(m, by=c("Nature", "Initial_SubjectiveDiscrimination", "Topic")) 
pred_conf$Topic <- paste0("MIST (", pred_conf$Topic, ")")
pred_conf$Probability <- pred_conf$Proportion
pred_conf$Proportion <- NULL
  
# MOCRI
m <- glmmTMB(
  Manipulative ~ Nature / Initial_SubjectiveDiscrimination + (Nature|Participant) + (1|Item),
  data = dfmocri[!is.na(dfmocri$Initial_SubjectiveDiscrimination) & dfmocri$Condition == "Pre-test",], 
  family = "binomial")
    
contrasts_mocri <- estimate_contrasts(m, contrast = "Initial_SubjectiveDiscrimination", by = c("Nature"))
contrasts_mocri$Topic <- "MOCRI"

pred_mocri <- estimate_means(m, by=c("Nature", "Initial_SubjectiveDiscrimination")) 
pred_mocri$Topic <- "MOCRI"


p <- rbind(pred_mist, pred_mocri) |> 
    ggplot(aes(x=Initial_SubjectiveDiscrimination, y=Probability)) +
    geom_pointrange(aes(ymin=CI_low, ymax=CI_high, color=Nature), position = position_dodge(width = 0.3)) +
    geom_line(aes(color=Nature, group=Nature), position = position_dodge(width = 0.3)) +
    scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
    scale_y_continuous(labels=scales::percent) + 
    theme_minimal() +
    facet_wrap(~Topic) +
    labs(x = "COVID Vaccination Status", y = "Probability of Believing an Item Fake/Manipulative") +
    theme(strip.background = element_rect(fill="lightgrey", color="white"),
          axis.title.x = element_text(face="bold"))
p

table <- rbind(contrasts_mist, contrasts_conf, contrasts_mocri) |> 
  select(-SE, -z) |> 
  insight::format_table() |> 
  gt::gt() |> 
  gt::data_color(columns = "Nature", palette = c("#F44336", "#4CAF50")) |> 
  gt::data_color(columns = "p", fn = function(x) ifelse(x < .01, "orange", ifelse(x < .05, "gold", "white")))
```

### Vaccination Status

```{r}
# MIST 
m <- glmmTMB(
  Belief ~ Nature / COVID_Vaccination * Topic + (Nature|Participant) + (1|Item),
  data = dfmist[!is.na(dfmist$COVID_Vaccination) & dfmist$Condition == "Pre-test",],
  family = "binomial"
)

contrasts_mist <- estimate_contrasts(m, contrast = "COVID_Vaccination", by = c("Nature", "Topic"))

pred_mist <- estimate_means(m, by=c("Nature", "COVID_Vaccination", "Topic")) 
pred_mist$Topic <- paste0("MIST (", pred_mist$Topic, ")")
  


m <- glmmTMB(
  Confidence ~ Nature / COVID_Vaccination * Topic + (Nature|Participant) + (1|Item),
  data = dfmist[!is.na(dfmist$COVID_Vaccination) & dfmist$Condition == "Pre-test" & dfmist$Correct == 1,],
  family = ordbeta()
)

contrasts_conf <- estimate_contrasts(m, contrast = "COVID_Vaccination", by = c("Nature", "Topic"))
contrasts_conf$Topic <- paste0("Confidence: ", contrasts_conf$Topic, "")

pred_conf <- estimate_means(m, by=c("Nature", "COVID_Vaccination", "Topic")) 
pred_conf$Topic <- paste0("MIST (", pred_conf$Topic, ")")
pred_conf$Probability <- pred_conf$Proportion
pred_conf$Proportion <- NULL
  
# MOCRI
m <- glmmTMB(
  Manipulative ~ Nature / COVID_Vaccination + (Nature|Participant) + (1|Item),
  data = dfmocri[!is.na(dfmocri$COVID_Vaccination) & dfmocri$Condition == "Pre-test",], 
  family = "binomial")
    
contrasts_mocri <- estimate_contrasts(m, contrast = "COVID_Vaccination", by = c("Nature"))
contrasts_mocri$Topic <- "MOCRI"

pred_mocri <- estimate_means(m, by=c("Nature", "COVID_Vaccination")) 
pred_mocri$Topic <- "MOCRI"


p <- rbind(pred_mist, pred_mocri) |> 
    ggplot(aes(x=COVID_Vaccination, y=Probability)) +
    geom_pointrange(aes(ymin=CI_low, ymax=CI_high, color=Nature), position = position_dodge(width = 0.3)) +
    geom_line(aes(color=Nature, group=Nature), position = position_dodge(width = 0.3)) +
    scale_color_manual(values=c("Fake"="#F44336", "Real"="#4CAF50", "Manipulative"="#F44336", "NonManipulative"="#4CAF50"), 
                       breaks = c("Fake", "Real"), 
                       labels = c("Fake / Manipulative", "Real / Non-Manipulative")) +
    scale_y_continuous(labels=scales::percent) + 
    theme_minimal() +
    facet_wrap(~Topic) +
    labs(x = "COVID Vaccination Status", y = "Probability of Believing an Item Fake/Manipulative") +
    theme(strip.background = element_rect(fill="lightgrey", color="white"),
          axis.title.x = element_text(face="bold"))
  

table <- rbind(contrasts_mist, contrasts_conf, contrasts_mocri) |> 
  select(-SE, -z) |> 
  insight::format_table() |> 
  gt::gt() |> 
  gt::data_color(columns = "Nature", palette = c("#F44336", "#4CAF50")) |> 
  gt::data_color(columns = "p", fn = function(x) ifelse(x < .01, "orange", ifelse(x < .05, "gold", "white")))
```


:::


## Figures

```{r}
#| fig-width: 10
#| fig-height: 13

((p1 + labs(title = "A. Effect Intervention")) /
  (rez_vsa$p + labs(x = "Authoritarianism", title = "B. Effect of Individual Characteristics")) /
  (rez_gcbsg$p + labs(x = "Conspiracy Beliefs") + theme(strip.text = element_blank())) /
  (rez_neuro$p + labs(x = "Neuroticism") + theme(strip.text = element_blank()))) +
  plot_layout(heights = c(1, 0.4, 0.4, 0.4), guides = "collect", axis_titles = "collect")
```

```{r}
#| fig-width: 10
#| fig-height: 7

p_mist2 
#   rez_vsa$p_conf /
#   rez_gcbs2$p_conf
```
